requirements.txt
Scrapy==1.5.0
boto3==1.5.29
txAWS==0.5.0
Pillow
fuzzywuzzy[speedup]
mock
pytest
scrapy-splash
raven
python-dateutil
pyopenssl==17.4.0
cryptography==1.9
